Define our Game data structure.
There are a lot of ways that we could have defined this data structure.

One of the decisions that I took here is that I want my data types to
build the `ubiquitous language` of the `tic tac toe` domain.

For example, we created a new data type called `Square` which is
isomorphic to `Maybe`. The benefits to this is that someone with little
programming knowledge can follow along and we are using the ubiquitous
language of Tic Tac Toe.

The downside to this that we lose out on all the infrastructure that has
been built for `Maybe`, including typeclass (`Functor`, `Applicative`,
`Monad`). Then again, should our `Square` really be a monad?

One way to have the best of both worlds is to use a `newtype` for our
`Square` type, and then use the `GeneralizedNewtypeDeriving` language
pragma.