Create a new data type ThreePeat.
The ThreePeat data type ensures that we have a three-element list. The
reason why we want to create a new data type is that `Row` and `Board`
are essentially equivalent. By abstracting this out into a new data
type, and making `Row` and `Board` newtypes around this data type, we
can create functions that manipulate `ThreePeat` once, instead of
creating equivalent functions for both `Row` and `Board`.