Implement ability to get current player.
We want to add a little type safety around this. `getCurrentPlayer` is
incorrect, because it is possible for the game to be over. In this case,
there shouldn't be a current player, but `getCurrentPlayer` will still
return one.

We could [expand the range][expand] of `getCurrentPlayer` so that the
final result is `Maybe Player`, but I like the idea that, instead of
doing this, we want to [push type safety bac][push], and make sure that
`getCurrentPlayer` can only be called with a game that is not already
won or tied. How do we do that?

[push]: http://www.parsonsmatt.org/2017/10/11/type_safety_back_and_forth.html
[expand]: http://www.parsonsmatt.org/2018/10/02/small_types.html